Synchronization point integrating all features into develop.

See merge request exadigit/raps!86

+++ raps/power.py
@@ -15,6 +15,7 @@ import numpy as np
 import pandas as pd
 import uncertainties as uf
 from .utils import linear_to_3d_index
+from .validators import recompute_power
 
 
 def custom_str_uncertainties(self):
@@ -271,7 +272,7 @@ class PowerManager:
         gpu_util = np.asarray(gpu_util)
         net_util = np.asarray(net_util)
         job_lengths = np.array([len(job) for job in scheduled_nodes])
-        flattened_nodes = np.concatenate(scheduled_nodes, axis=0)
+        flattened_nodes = np.concatenate(scheduled_nodes, axis=0).astype(np.int64)
 
         cpu_util_flat = np.repeat(cpu_util, job_lengths)
         gpu_util_flat = np.repeat(gpu_util, job_lengths)
